Added two more modules, ranger-http-server-bundle and ranger-zk-server-bundle. For those using the same ShardInfo in their code to write a server, all they need to now do is write their dummy app interface and write it for packaging only, taking away the boilerplate they may have to replicate
