+++ pom.xml
@@ -25,6 +25,8 @@
         <module>ranger-zk-server</module>
         <module>ranger-http-server</module>
         <module>ranger-http-model</module>
+        <module>ranger-http-server-bundle</module>
+        <module>ranger-zk-server-bundle</module>
     </modules>
 
     <distributionManagement>
