+++ ranger-http-server-bundle/src/main/java/io/appform/ranger/http/server/bundle/healthcheck/RangerHttpHealthCheck.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2015 Flipkart Internet Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.appform.ranger.http.server.bundle.healthcheck;
+
+import com.codahale.metrics.health.HealthCheck;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import javax.inject.Singleton;
+
+@EqualsAndHashCode(callSuper = true)
+@Singleton
+@Data
+public class RangerHttpHealthCheck extends HealthCheck {
+    @Override
+    protected Result check() {
+        return HealthCheck.Result.healthy("Healthy");
+    }
+}
