+++ ranger-zk-server/pom.xml
@@ -11,41 +11,11 @@
 
     <artifactId>ranger-zk-server</artifactId>
 
-    <properties>
-        <dropwizard.version>2.0.23</dropwizard.version>
-    </properties>
-
     <dependencies>
         <dependency>
             <groupId>io.appform.ranger</groupId>
-            <artifactId>ranger-zk-client</artifactId>
+            <artifactId>ranger-zk-server-bundle</artifactId>
             <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.appform.ranger</groupId>
-            <artifactId>ranger-server-bundle</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.appform.ranger</groupId>
-            <artifactId>ranger-server-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.dropwizard</groupId>
-            <artifactId>dropwizard-core</artifactId>
-            <version>${dropwizard.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-            <version>${mockito.version}</version>
-            <scope>test</scope>
         </dependency>
     </dependencies>
